move otel config stuff
