+++ shared/observability/src/airflow_shared/observability/traces/__init__.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,6 +17,64 @@
 # under the License.
 from __future__ import annotations
 
-TRACEPARENT = "traceparent"
-TRACESTATE = "tracestate"
-NO_TRACE_ID = 1
+import logging
+from contextlib import contextmanager
+
+from opentelemetry import context, trace
+from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import (
+    BatchSpanProcessor,
+)
+from opentelemetry.sdk.trace.id_generator import RandomIdGenerator
+
+log = logging.getLogger(__name__)
+
+OVERRIDE_SPAN_ID_KEY = context.create_key("override_span_id")
+OVERRIDE_TRACE_ID_KEY = context.create_key("override_trace_id")
+
+
+class OverrideableRandomIdGenerator(RandomIdGenerator):
+    """Lets you override the span id."""
+
+    def generate_span_id(self):
+        override = context.get_value(OVERRIDE_SPAN_ID_KEY)
+        if override is not None:
+            context.attach(context.set_value(OVERRIDE_SPAN_ID_KEY, None))
+            return override
+        return super().generate_span_id()
+
+    def generate_trace_id(self):
+        override = context.get_value(OVERRIDE_TRACE_ID_KEY)
+        if override is not None:
+            context.attach(context.set_value(OVERRIDE_TRACE_ID_KEY, None))
+            return override
+        return super().generate_trace_id()
+
+
+@contextmanager
+def override_ids(trace_id, span_id, ctx=None):
+    ctx = context.set_value(OVERRIDE_TRACE_ID_KEY, trace_id, context=ctx)
+    ctx = context.set_value(OVERRIDE_SPAN_ID_KEY, span_id, context=ctx)
+    token = context.attach(ctx)
+    try:
+        yield
+    finally:
+        context.detach(token)
+
+
+# port = conf.getint("traces", "otel_port", fallback=None)
+# host = conf.get("traces", "otel_host", fallback=None)
+# ssl_active = conf.getboolean("traces", "otel_ssl_active", fallback=False)
+# otel_service = conf.get("traces", "otel_service", fallback=None)
+
+# resource = Resource.create(attributes={SERVICE_NAME: otel_service})
+# otel_env_config = load_traces_env_config()
+
+
+def configure_otel():
+    provider = TracerProvider(id_generator=OverrideableRandomIdGenerator())
+    exporter = OTLPSpanExporter()
+    span_processor = BatchSpanProcessor(exporter)
+    provider.add_span_processor(span_processor)
+    trace.set_tracer_provider(provider)
